Extract result-related structures to dedicated result.go file
- Create result.go with all result-related type definitions
- Move StatusType, StepResult, JobResult, Result structs from printer.go
- Move StepRepeatCounter from step.go to result.go
- Move NewResult() function and AddStepResult() method
- Remove unused sync import from printer.go
- Improve code organization and separation of concerns
